+++ client/src/components/Editor/collaboration.ts
@@ -1,7 +1,7 @@
 import * as Y from 'yjs';
 import { WebrtcProvider } from 'y-webrtc';
 
-export const ROOM_NAME = 'pair-programming-terminal-room';
+export const ROOM_NAME = 'redocly-pair-programming';
 
 export const ydoc = new Y.Doc();
 export const provider = new WebrtcProvider(ROOM_NAME, ydoc);
